Seer beta banner copy changes
